Bug 1847469 - Part 15: Use uint32_t for line and column in debugger as a preparation to use column number types. r=iain
Differential Revision: https://phabricator.services.mozilla.com/D185753

UltraBlame original commit: 64a80f31ee5e6d9700aa3cc7fa76966275249df2

+++ js/src/vm/BytecodeUtil-inl.h
@@ -162,8 +162,8 @@ class BytecodeRangeWithPosition : private BytecodeRange {
     }
   }
 
-  size_t frontLineNumber() const { return lineno; }
-  size_t frontColumnNumber() const { return column; }
+  uint32_t frontLineNumber() const { return lineno; }
+  uint32_t frontColumnNumber() const { return column; }
 
   
   
@@ -226,13 +226,13 @@ class BytecodeRangeWithPosition : private BytecodeRange {
     isEntryPoint = lastLinePC == frontPC();
   }
 
-  size_t initialLine;
+  uint32_t initialLine;
 
   
-  size_t lineno;
+  uint32_t lineno;
 
   
-  size_t column;
+  uint32_t column;
 
   const SrcNote* sn;
   jsbytecode* snpc;
